+++ src/components/mod.rs
@@ -6,7 +6,6 @@ use bevy_ecs_ldtk::{LdtkEntity, LdtkIntCell};
 use bevy_xpbd_2d::components::{Collider, CollisionLayers, Position};
 use bevy_xpbd_2d::math::Vector2;
 use bevy_xpbd_2d::prelude::{RigidBody, Rotation};
-use big_brain::prelude::{ActionBuilder, ScorerBuilder};
 use crate::{Layer, METERS_PER_PIXEL};
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Default, Component)]
@@ -71,46 +70,7 @@ impl Default for PlayerBundle {
 pub struct Player {}
 
 #[derive(Component, Clone)]
-pub struct Boid {}
-
-#[derive(Component, Clone)]
 pub struct Prey {}
-
-#[derive(Reflect)]
-#[derive(Copy, Clone, Debug, Component)]
-pub struct BoidStuff {
-    pub cohesion_boids: i32,
-    pub separation_boids: i32,
-    pub flock_center: Vector2,
-    pub separation_vector: Vector2,
-    pub separation_distance: f32,
-    pub cohesion_distance: f32,
-    pub separation_factor: f32,
-    pub cohesion_factor: f32,
-    pub alignment_direction: Vector2,
-    pub alignment_distance: f32,
-    pub alignment_factor: f32,
-    pub alignment_boids: i32,
-}
-
-impl Default for BoidStuff {
-    fn default() -> Self {
-        Self {
-            cohesion_boids: 0,
-            separation_boids: 0,
-            alignment_boids: 0,
-            flock_center: Vector2::ZERO,
-            separation_vector: Vector2::ZERO,
-            alignment_direction: Vector2::ZERO,
-            separation_distance: 25.0,
-            cohesion_distance: 100.0,
-            alignment_distance: 75.0,
-            separation_factor: 0.5,
-            cohesion_factor: 0.5,
-            alignment_factor: 0.7,
-        }
-    }
-}
 
 #[derive(Component)]
 pub struct AimLine {}
@@ -136,24 +96,6 @@ impl Default for DirectionControl {
             aim_rotation: Rotation::default(),
             aim_degrees: 0.0,
             mouse_position: Vector2::ZERO,
-            force_scale: 10.0,
-        }
-    }
-}
-
-#[derive(Reflect)]
-#[derive(Copy, Clone, Debug, Component)]
-pub struct BoidDirection {
-    pub direction: Vector2,
-    pub up: Vector2,
-    pub force_scale: f32,
-}
-
-impl Default for BoidDirection {
-    fn default() -> Self {
-        Self {
-            direction: Vector2::ZERO,
-            up: Vector2::Y,
             force_scale: 10.0,
         }
     }
@@ -204,25 +146,5 @@ pub struct IntCell {
 
 #[derive(Component, Debug, Reflect, Clone)]
 pub struct AAName(pub String);
-
-#[derive(Component, Debug)]
-pub struct Hunger {
-    pub per_second: f32,
-    pub hunger: f32,
-}
-
-#[derive(Clone, Component, Debug, ActionBuilder)]
-pub struct Hunt {
-    pub until: f32
-}
-
-impl Hunger {
-    pub fn new(hunger: f32, per_second: f32) -> Self {
-        Self { hunger, per_second }
-    }
-}
-
-#[derive(Clone, Component, Debug, ScorerBuilder)]
-pub struct Hungry;
